requirements.txt
playwright>=1.49.0
pyyaml>=6.0.2
